requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
Django==6.0
flet==0.28.3
flet-desktop==0.28.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
packaging==24.2
pillow==12.1.0
psycopg2-binary==2.9.11
pymodbus==3.11.4
pyModbusTCP==0.1.10
python-dotenv==1.2.1
qrcode==8.2
repath==0.9.0
six==1.17.0
sqlparse==0.5.4
tk==0.1.0
tzdata==2025.2
waitress==3.0.2
